# MusicAPI/openapi.yml
openapi: '3.0.3'
info:
  title: Music Social Media API
  description: Rest API for a social media platform with ML-based Content/Collaborative/Hybrid User Specific Recommendations
  version: '1.0'
servers:
  - url: https://mt241063-10974.node.ustp.cloud
    description: Deployment
  - url: http://localhost:3000/
    description: Local Testing
paths:
...
  /users/follow/{username}:
    post:
      summary: "Triggered when a user wants to follow another user"
      tags: [ UserActions ]
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Followed successfully"

        "400":
          description: "Already Following"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                AlreadyFollowing:
                  value:
                    message: "You are already following this user"
                    code: "ALREADY_FOLLOWING"
                CannotFollowSelf:
                  value:
                    message: "You cannot follow yourself"
                    code: "CANNOT_FOLLOW_SELF"
                MissingFields:
                  value:
                    message: "Missing required fields"
                    code: "MISSING_FIELDS"


        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "You need to be logged in in order to follow another user."
                code: "NOT_LOGGED_IN"

        "404":
          description: "User does not exist"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "The user you are trying to follow does not exist"
                code: "USER_DOES_NOT_EXIST"

    delete:
      summary: "Triggered when a user wants to unfollow another user"
      tags: [ UserActions ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Unfollowed successfully"

        400":
          description: "Already Following"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:


                MissingFields:
                  value:
                    message: "Missing required fields"
                    code: "MISSING_FIELDS"
                NotFollowing:
                  value:
                    message: "You´re not following this user."
                    code: "NOT_FOLLOWED"
                CannotUnfollowSelf:
                  value:
                    message: "You cannot unfollow yourself"
                    code: "CANNOT_UNFOLLOW_SELF"

        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "You need to be logged in in order to unfollow another user."
                code: "NOT_LOGGED_IN"

        "404":
          description: "User does not exits"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "The user you are trying to unfollow does not exist"
                code: "USER_DOES_NOT_EXIST"
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          example: "Wrong credentials"
        code:
          type: string
          example: "INVALID_CREDENTIALS"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT